Brushing game.
Imagine a toy car with a brush in front of it circulating in a room. The car is controlled by python application using control flows, loops and functions. Brush in the car can be in two
different positions: above or below. When the brush is down, it paints the places where the vehicle goes and does nothing when the brush is up. That program is implemented with python.